Revert "Revert "KBC-50 Add tests for import/export on azure blob storage""
